backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
pydantic==2.9.2
pydantic-settings==2.6.0
sqlalchemy==2.0.36
python-dotenv==1.0.1
google-generativeai==0.8.3
python-multipart==0.0.12
